moved confirmation dialog component in common folder
