+++ src/app/components/core/control-dispatcher/control-dispatcher.module.ts
@@ -6,8 +6,10 @@ import { RouterModule, Routes } from '@angular/router';
 import { TreeModule } from '@circlon/angular-tree-component';
 import { NgSelectModule } from '@ng-select/ng-select';
 import { DirectivesModule } from '../../../directives/directives.module';
-import { AdminGuard, AuthGuard, AuthTypeGuard } from '../../../guards';
+import { AuthTypeGuard } from '../../../guards';
 import { PipesModule } from '../../../pipes/pipes.module';
+import { SharedModule } from '../../../shared.module';
+import { AddControlAclComponent } from './add-control-acl/add-control-acl.component';
 import { AddControlScriptComponent } from './add-control-script/add-control-script.component';
 import { AddConfigureComponent } from './add-control-script/add-step/add-configure/add-configure.component';
 import { AddDelayComponent } from './add-control-script/add-step/add-delay/add-delay.component';
@@ -17,13 +19,11 @@ import { AddStepConditionComponent } from './add-control-script/add-step/add-ste
 import { AddStepValueComponent } from './add-control-script/add-step/add-step-value/add-step-value.component';
 import { AddStepComponent } from './add-control-script/add-step/add-step.component';
 import { AddWriteComponent } from './add-control-script/add-step/add-write/add-write.component';
-import { ConfirmationDialogComponent } from './confirmation-dialog/confirmation-dialog.component';
-import { ControlScriptsListComponent } from './list-control-dispatcher/control-scripts-list/control-scripts-list.component';
-import { ListControlDispatcherComponent } from './list-control-dispatcher/list-control-dispatcher.component';
-import { AclListComponent } from './list-control-dispatcher/acl-list/acl-list.component';
-import { AddControlAclComponent } from './add-control-acl/add-control-acl.component';
-import { ControlTasksListComponent } from './list-control-dispatcher/control-tasks-list/control-tasks-list.component';
 import { ControlScheduleTaskDetailsComponent } from './control-schedule-task/control-schedule-task-details.component';
+import { AclListComponent } from './list-control-dispatcher/acl-list/acl-list.component';
+import { ControlScriptsListComponent } from './list-control-dispatcher/control-scripts-list/control-scripts-list.component';
+import { ControlTasksListComponent } from './list-control-dispatcher/control-tasks-list/control-tasks-list.component';
+import { ListControlDispatcherComponent } from './list-control-dispatcher/list-control-dispatcher.component';
 
 const routes: Routes = [
 
@@ -73,7 +73,6 @@ const routes: Routes = [
     AddScriptComponent,
     AddConfigureComponent,
     AddDelayComponent,
-    ConfirmationDialogComponent,
     AclListComponent,
     AddControlAclComponent,
     ControlTasksListComponent,
@@ -88,6 +87,7 @@ const routes: Routes = [
     TreeModule,
     NgSelectModule,
     DirectivesModule,
+    SharedModule,
     RouterModule.forChild(routes)
   ],
   providers: [AuthTypeGuard]
