+++ src/app/components/core/control-dispatcher/list-control-dispatcher/control-scripts-list/control-scripts-list.component.ts
@@ -3,8 +3,8 @@ import { Subscription } from 'rxjs';
 import { orderBy } from 'lodash';
 import { AlertService, ProgressBarService, SharedService } from '../../../../../services';
 import { ControlDispatcherService } from '../../../../../services/control-dispatcher.service';
-import { ConfirmationDialogComponent } from '../../confirmation-dialog/confirmation-dialog.component';
-import { DialogService } from '../../confirmation-dialog/dialog.service';
+import { ConfirmationDialogComponent } from '../../../../common/confirmation-dialog/confirmation-dialog.component';
+import { DialogService } from '../../../../common/confirmation-dialog/dialog.service';
 
 @Component({
   selector: 'app-control-scripts-list',
