+++ src/app/shared.module.ts
@@ -1,25 +1,50 @@
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { NgSelectModule } from '@ng-select/ng-select';
 import { CodemirrorModule } from '@ctrl/ngx-codemirror';
+import { NgSelectModule } from '@ng-select/ng-select';
 
+import { ConfirmationDialogComponent } from './components/common/confirmation-dialog/confirmation-dialog.component';
+import { RangeSliderComponent } from './components/common/range-slider/range-slider.component';
+import { TimeDropdownComponent } from './components/common/time-dropdown/time-dropdown.component';
+import { ConfigChildrenComponent } from './components/core/configuration-manager/config-children/config-children.component';
 import {
-  ViewConfigItemComponent,
+  ViewConfigItemComponent
 } from './components/core/configuration-manager/view-config-item/view-config-item.component';
 import {
-  ViewLogsComponent,
+  ViewLogsComponent
 } from './components/core/packages-log/view-logs/view-logs.component';
+import { PluginModalComponent } from './components/core/plugin-modal/plugin-modal.component';
 import { DirectivesModule } from './directives/directives.module';
 import { PipesModule } from './pipes/pipes.module';
-import { ConfigChildrenComponent } from './components/core/configuration-manager/config-children/config-children.component';
-import { PluginModalComponent } from './components/core/plugin-modal/plugin-modal.component';
-import { TimeDropdownComponent } from './components/common/time-dropdown/time-dropdown.component';
-import { RangeSliderComponent } from './components/common/range-slider/range-slider.component';
 
 @NgModule({
-  imports: [CommonModule, CodemirrorModule, PipesModule, FormsModule, DirectivesModule, NgSelectModule, ReactiveFormsModule],
-  declarations: [ViewConfigItemComponent, ConfigChildrenComponent, PluginModalComponent, ViewLogsComponent, TimeDropdownComponent, RangeSliderComponent],
-  exports: [ViewConfigItemComponent, ConfigChildrenComponent, PluginModalComponent, ViewLogsComponent, TimeDropdownComponent, RangeSliderComponent]
+  imports: [
+    CommonModule,
+    CodemirrorModule,
+    PipesModule,
+    FormsModule,
+    DirectivesModule,
+    NgSelectModule,
+    ReactiveFormsModule
+  ],
+  declarations: [
+    ViewConfigItemComponent,
+    ConfigChildrenComponent,
+    PluginModalComponent,
+    ViewLogsComponent,
+    TimeDropdownComponent,
+    RangeSliderComponent,
+    ConfirmationDialogComponent
+  ],
+  exports: [
+    ViewConfigItemComponent,
+    ConfigChildrenComponent,
+    PluginModalComponent,
+    ViewLogsComponent,
+    TimeDropdownComponent,
+    RangeSliderComponent,
+    ConfirmationDialogComponent
+  ]
 })
 export class SharedModule { }
